Move storage settings handling code to a module
