+++ rust/agama-lib/src/storage/proxies.rs
@@ -1,0 +1,106 @@
+//! D-Bus interface proxies for: `org.opensuse.Agama.Storage1.*`
+//!
+//! This code was generated by `zbus-xmlgen` `3.1.0` from DBus introspection data.`.
+use zbus::dbus_proxy;
+
+#[dbus_proxy(
+    interface = "org.opensuse.Agama1.Validation",
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1"
+)]
+trait Validation {
+    /// Errors property
+    #[dbus_proxy(property)]
+    fn errors(&self) -> zbus::Result<Vec<String>>;
+
+    /// Valid property
+    #[dbus_proxy(property)]
+    fn valid(&self) -> zbus::Result<bool>;
+}
+
+#[dbus_proxy(
+    interface = "org.opensuse.Agama.Storage1",
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1"
+)]
+trait Storage1 {
+    /// Finish method
+    fn finish(&self) -> zbus::Result<()>;
+
+    /// Install method
+    fn install(&self) -> zbus::Result<()>;
+
+    /// Probe method
+    fn probe(&self) -> zbus::Result<()>;
+
+    /// DeprecatedSystem property
+    #[dbus_proxy(property)]
+    fn deprecated_system(&self) -> zbus::Result<bool>;
+}
+
+#[dbus_proxy(
+    interface = "org.opensuse.Agama.Storage1.Proposal.Calculator",
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1"
+)]
+trait Calculator {
+    /// Calculate method
+    fn calculate(
+        &self,
+        settings: std::collections::HashMap<&str, zbus::zvariant::Value<'_>>,
+    ) -> zbus::Result<u32>;
+
+    /// AvailableDevices property
+    #[dbus_proxy(property)]
+    fn available_devices(
+        &self,
+    ) -> zbus::Result<
+        Vec<(
+            String,
+            String,
+            std::collections::HashMap<String, zbus::zvariant::OwnedValue>,
+        )>,
+    >;
+
+    /// Result property
+    #[dbus_proxy(property)]
+    fn result(&self) -> zbus::Result<zbus::zvariant::OwnedObjectPath>;
+
+    /// VolumeTemplates property
+    #[dbus_proxy(property)]
+    fn volume_templates(
+        &self,
+    ) -> zbus::Result<Vec<std::collections::HashMap<String, zbus::zvariant::OwnedValue>>>;
+}
+
+#[dbus_proxy(
+    interface = "org.opensuse.Agama.Storage1.Proposal",
+    default_service = "org.opensuse.Agama.Storage1",
+    default_path = "/org/opensuse/Agama/Storage1/Proposal"
+)]
+trait StorageProposal {
+    /// Actions property
+    #[dbus_proxy(property)]
+    fn actions(
+        &self,
+    ) -> zbus::Result<Vec<std::collections::HashMap<String, zbus::zvariant::OwnedValue>>>;
+
+    /// CandidateDevices property
+    #[dbus_proxy(property)]
+    fn candidate_devices(&self) -> zbus::Result<Vec<String>>;
+
+    /// EncryptionPassword property
+    #[dbus_proxy(property)]
+    fn encryption_password(&self) -> zbus::Result<String>;
+
+    /// LVM property
+    #[dbus_proxy(property, name = "LVM")]
+    fn lvm(&self) -> zbus::Result<bool>;
+
+    /// Volumes property
+    #[dbus_proxy(property)]
+    fn volumes(
+        &self,
+    ) -> zbus::Result<Vec<std::collections::HashMap<String, zbus::zvariant::OwnedValue>>>;
+}
+
